rotation X,Y,Z qubit gates
implementation of X,Y,Z rotation gates, cleanup for tf_qcs_real_number

+++ include/qcs_misc.h
@@ -30,8 +30,8 @@
 
 
 //#define tf_qcs_real_number double
-typedef float tf_qcs_real_number;
-//typedef double tf_qcs_real_number ;
+//typedef float tf_qcs_real_number;
+typedef double tf_qcs_real_number;
 
 #define QCS_ACCURACY -7
 #define QCS_EPS 0.000001
